Update README.md
This is a great definition of technical writing. The cultural iceberg is a useful metaphor for thinking about different audiences. Minor cultural differences can have a significant impact on communication effectiveness in critical situations, so it's always important to consider what's "beneath the waterline."

Your Markdown syntax is fine, but you posted this as a comment to the commit instead of to the file itself. Click the little pencil to edit files directly. 

I've moved your post over to the README.md file in a new branch of your repository. Accept the changes and merge them back into your "master" branch. I'll use the "adavis46-feedback" branch to give you comments and suggestions and you can decide whether or not to accept them back into the primary branch. Hopefully this will start to visualize how Git can be used for version control.